requirements.txt
Flask
psycopg2
flask_cors
load_dotenv